Add vanity key miner functionality
This commit adds functionalities to the `vainstr` vanity key miner. It introduces `go.mod` and `go.sum` dependencies, modifies `.gitignore`, `README.md` and it creates `main.go` with the logic for key pair generation along with vanity mining. According to the set arguments, it attempts to find a string within a generated public key. The Position argument specifies where in the key the string should be found. The number of threads for mining can also be specified.